requirements.txt
asgiref==3.6.0
Babel==2.12.1
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
